fix: call CKAN dataset-page in the background to trigger default-view creation
CKAN does not have an API hook to trigger this, and calling the dataset or resource via API (e.g. package_show or resource_show) will not call the `before_dataset_view` Interface-functions as the context is not `for_view`.

This is a hack, but as it is done on the server and we don't need the result, it should not interfere with anything on the frontend. If the backgroundURL can not be fetched, we certainly have other problems..